piano-hero/requirements.txt
# Piano Hero Cover - Зависимости Python

# Основные зависимости
numpy>=1.21.0
PyYAML>=6.0
basic-pitch>=0.1.0
pretty-midi>=0.2.9

# Обработка аудио
librosa>=0.9.0
soundfile>=0.10.0

# Обработка MIDI
mido>=1.2.0

# Утилиты
tqdm>=4.64.0
pathlib2>=2.3.0

# Дополнительные зависимости для обработки
scipy>=1.8.0
matplotlib>=3.5.0

# Зависимости для работы с TensorFlow (для Basic Pitch)
tensorflow>=2.8.0
tensorflow-io>=0.24.0

# Зависимости для работы с аудио
pydub>=0.25.0



